Refactor file structure (#24)
